+++ .gitignore
@@ -1,3 +1,3 @@
-am/__pycache__/**
+__pycache__
 am/parser.out
 am/parsetab.py
